refactor: update ShiftTable directory name
